Ataul: Código fracionado em várias partes por export/import;
